File structure improvements v1
