+++ humancompatible/detect/helpers/utils.py
