+++ humancompatible/detect/methods/__init__.py
@@ -1,4 +1,4 @@
-from .detect import detect_bias, detect_bias_csv, detect_bias_two_samples, prepare_dataset
+from ..detect import detect_bias, detect_bias_csv, detect_bias_two_samples, prepare_dataset
 
 __all__ = [
     "detect_bias",
