+++ humancompatible/detect/methods/l_inf/msd/msd.py
@@ -10,12 +10,11 @@ logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO, format="[%(levelname)s] %(message)s")
 
 
-def compute_MSD(
+def get_conjuncts_MSD(
     X_bin: np.ndarray[np.bool_],
     y_bin: np.ndarray[np.bool_],
-    time_limit: int = 600,
-    n_min: int = 0,
-    solver: str = "appsi_highs",
+    rule
+    
 ) -> Tuple[float, List[int]]:
     """Computes the Maximum Subgroup Discrepancy (MSD) to detect bias.
 
@@ -46,10 +45,7 @@ def compute_MSD(
               defines the subgroup with maximal discrepancy.
     """
 
-    mio = OneRule()
-    rule, _ = mio.find_rule(
-        X_bin, y_bin, n_min=n_min, time_limit=time_limit, solver_name=solver
-    )
+    
     subgroup_map = subgroup_map_from_conjuncts_binarized(rule, X_bin)
     MSD_val = evaluate_subgroup_discrepancy(subgroup_map, y_bin)
